Fix a couple of problems with the centos6 bootstrap process
rinse is not my best friend right now.  It installs *every* package in
its cache, even if those packages are now horribly out-of-date, which
tends to cause all manner of fun and games.  So, we turn caching off.
It also leaves the system in a horrendously broken state, with brokne
dependencies all over the place.  So, we've got to go in and manually
fix that shit up.